EOW project: add generate_low_stock_report to Invnetory class, refactored: old code

+++ Week 2/Day 6/package/inventory_manager/inventory_manager/src/utility.py
@@ -54,7 +54,7 @@ class ProductDetails:
     type: str
     price: float
     quantity: int
-    days_to_expire: float 
+    days_to_expire: int
     is_vegetarian: bool | str | None
     warranty_period_in_years: int
